🔨 Update | 인벤토리 - 거래 시스템 일반화 코드 추출
