+++ Assets/Scripts/Inventory/UI/UI_InvenSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d91f61993e9c7b843a52d0fdf03399be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
