--- Assets/_Assets/Cainos/Pixel Art Monster - Dungeon/Patch - URP 2D Lit.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6c1976b53e0a37a4e9de1feaf72bd45e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
